Fix timer bug and cache bug
