+++ pkg/async/service.go
@@ -107,12 +107,11 @@ func (lat *TaskService) daemon(ctx context.Context) {
 	util.SafeCloseChan(lat.ready)
 	ticker := time.NewTicker(removeExecutorInterval)
 	max := 0
+	timer := time.NewTimer(executeInterval)
+	defer timer.Stop()
 	for {
-		timer := time.NewTimer(executeInterval)
 		select {
 		case <-ctx.Done():
-			timer.Stop()
-
 			util.Logger().Debugf("daemon thread exited for TaskService is stopped")
 			return
 		case <-timer.C:
@@ -127,8 +126,10 @@ func (lat *TaskService) daemon(ctx context.Context) {
 			for _, s := range slice {
 				s.Execute() // non-blocked
 			}
+
+			timer.Reset(executeInterval)
 		case <-ticker.C:
-			timer.Stop()
+			util.ResetTimer(timer, executeInterval)
 
 			lat.lock.RLock()
 			l := len(lat.executors)
