+++ pkg/cache/cache_test.go
@@ -17,6 +17,7 @@
 package cache
 
 import (
+	"fmt"
 	"golang.org/x/net/context"
 	"testing"
 	"time"
@@ -31,12 +32,18 @@ func (l *level1) Name(ctx context.Context) string {
 
 func (l *level1) Init(ctx context.Context, parent *Node) (node *Node, err error) {
 	p := l.Name(ctx)
-	if p == "1" {
-		node = NewNode()
-		node.Cache.Set("a", "a")
-		node.Cache.Set("b", "b")
-		node.Cache.Set("c", "c")
+	if p == "err" {
+		return nil, fmt.Errorf("wrong logic")
 	}
+
+	if p == "null" {
+		return
+	}
+
+	node = NewNode()
+	node.Cache.Set("a", "a")
+	node.Cache.Set("b", "b")
+	node.Cache.Set("c", "c")
 	return
 }
 
@@ -54,6 +61,14 @@ func (l *level2) Init(ctx context.Context, parent *Node) (node *Node, err error)
 	}
 
 	p := l.Name(ctx)
+	if p == "err" {
+		return nil, fmt.Errorf("wrong logic")
+	}
+
+	if p == "null" {
+		return
+	}
+
 	node = NewNode()
 	if len(l.changed) == 0 {
 		i := parent.Cache.Get(p)
@@ -71,13 +86,35 @@ func TestTree_Get(t *testing.T) {
 	if node != nil || err != nil {
 		t.Fatalf("TestTree_Get failed")
 	}
+	tree.Remove(context.Background())
 
 	l2 := &level2{}
 	tree.AddFilter(&level1{}, l2)
 
-	ctx := context.WithValue(context.WithValue(context.Background(), "key1", ""), "key2", "")
+	ctx := context.WithValue(context.WithValue(context.Background(), "key1", "err"), "key2", "")
 	node, err = tree.Get(ctx)
+	if node != nil || err == nil {
+		t.Fatalf("TestTree_Get failed")
+	}
+
+	node, err = tree.Get(context.WithValue(context.WithValue(ctx, "key1", "null"), "key2", ""))
 	if node != nil || err != nil {
+		t.Fatalf("TestTree_Get failed")
+	}
+
+	node, err = tree.Get(context.WithValue(context.WithValue(ctx, "key1", "1"), "key2", "err"))
+	node, err = tree.Get(ctx)
+	if node != nil || err == nil {
+		t.Fatalf("TestTree_Get failed")
+	}
+
+	node, err = tree.Get(context.WithValue(context.WithValue(ctx, "key1", "1"), "key2", "null"))
+	if node != nil || err != nil {
+		t.Fatalf("TestTree_Get failed")
+	}
+
+	node, err = tree.Get(context.WithValue(context.WithValue(ctx, "key1", "1"), "key2", "err"))
+	if node != nil || err == nil {
 		t.Fatalf("TestTree_Get failed")
 	}
 
@@ -117,7 +154,8 @@ func TestTree_Get(t *testing.T) {
 	}
 
 	l2.changed = ""
-	node, err = tree.Simulate(context.WithValue(context.WithValue(ctx, "key1", "1"), "key2", "b"))
+	node, err = tree.Get(context.WithValue(context.WithValue(ctx, "key1", "1"), "key2", "b"),
+		Options().Temporary(true))
 	if node == nil || err != nil {
 		t.Fatalf("TestTree_Get failed")
 	}
@@ -125,7 +163,7 @@ func TestTree_Get(t *testing.T) {
 		t.Fatalf("TestTree_Get failed")
 	}
 
-	node, err = tree.Get(context.WithValue(ctx, "key1", "1"), Options().BeforeLevel(1))
+	node, err = tree.Get(context.WithValue(ctx, "key1", "1"), Options().ToLevel(1))
 	if node == nil || err != nil {
 		t.Fatalf("TestTree_Get failed")
 	}
