+++ pkg/cache/node.go
@@ -30,9 +30,7 @@ type Node struct {
 
 func (n *Node) ChildNodes() (nodes []*Node) {
 	n.Childs.ForEach(func(item util.MapItem) (next bool) {
-		if node := n.Tree.Nodes(n.Level+1, item.Key.(string)); node != nil {
-			nodes = append(nodes, node)
-		}
+		nodes = append(nodes, item.Value.(*Node))
 		return true
 	})
 	return
