+++ pkg/cache/tree.go
@@ -18,7 +18,6 @@ package cache
 
 import (
 	"errors"
-	"github.com/apache/incubator-servicecomb-service-center/pkg/util"
 	"github.com/karlseguin/ccache"
 	"golang.org/x/net/context"
 	"sync"
@@ -28,7 +27,7 @@ var errNilNode = errors.New("nil node")
 
 type Tree struct {
 	Config  *Config
-	nodes   []*ccache.Cache
+	roots   *ccache.Cache
 	filters []Filter
 	lock    sync.RWMutex
 }
@@ -37,7 +36,6 @@ func (t *Tree) AddFilter(fs ...Filter) *Tree {
 	t.lock.Lock()
 	for _, f := range fs {
 		t.filters = append(t.filters, f)
-		t.nodes = append(t.nodes, ccache.New(ccache.Configure().MaxSize(t.Config.MaxSize())))
 	}
 	t.lock.Unlock()
 	return t
@@ -49,12 +47,24 @@ func (t *Tree) Get(ctx context.Context, ops ...Option) (node *Node, err error) {
 		op = ops[0]
 	}
 
-	var parent *Node
-	for i := range t.filters {
+	var (
+		parent *Node
+		i      int
+	)
+
+	if !op.NoCache {
+		if parent, err = t.getOrCreateRoot(ctx); parent == nil {
+			return
+		}
+		i++
+		// parent may be a temp root in concurrent scene
+	}
+
+	for ; i < len(t.filters); i++ {
 		if op.Level > 0 && op.Level == i {
 			break
 		}
-		if parent, err = t.getOrCreateNode(ctx, i, parent); parent == nil || err != nil {
+		if parent, err = t.getOrCreateNode(ctx, i, parent); parent == nil {
 			break
 		}
 	}
@@ -67,29 +77,18 @@ func (t *Tree) Remove(ctx context.Context) {
 		return
 	}
 
-	t.remove(0, t.filters[0].Name(ctx))
+	t.roots.Delete(t.filters[0].Name(ctx))
 }
 
-func (t *Tree) remove(idx int, name string) {
-	if parent := t.Nodes(idx, name); parent != nil {
-		parent.Childs.ForEach(func(item util.MapItem) (next bool) {
-			t.remove(idx+1, item.Key.(string))
-			return true
-		})
-		t.nodes[parent.Level].Delete(parent.Name)
-	}
-}
-
-func (t *Tree) getOrCreateNode(ctx context.Context, idx int, parent *Node) (node *Node, err error) {
-	if len(t.filters) <= idx {
+func (t *Tree) getOrCreateRoot(ctx context.Context) (node *Node, err error) {
+	if len(t.filters) == 0 {
 		return
 	}
 
-	filter := t.filters[idx]
-	name := t.nodeFullName(filter.Name(ctx), parent)
-	item, err := t.nodes[idx].Fetch(name, t.Config.TTL(), func() (interface{}, error) {
-		// if node is miss or stale
-		node, err := t.createNode(ctx, idx, name, parent)
+	filter := t.filters[0]
+	name := filter.Name(ctx)
+	item, err := t.roots.Fetch(name, t.Config.TTL(), func() (interface{}, error) {
+		node, err := t.getOrCreateNode(ctx, 0, nil)
 		if err != nil {
 			return nil, err
 		}
@@ -107,22 +106,39 @@ func (t *Tree) getOrCreateNode(ctx context.Context, idx int, parent *Node) (node
 	return
 }
 
+func (t *Tree) getOrCreateNode(ctx context.Context, idx int, parent *Node) (node *Node, err error) {
+	filter := t.filters[idx]
+	name := t.nodeFullName(filter.Name(ctx), parent)
+
+	if parent == nil {
+		// new a temp node
+		return t.createNode(ctx, idx, name, parent)
+	}
+
+	item, err := parent.Childs.Fetch(name, func() (interface{}, error) {
+		node, err := t.createNode(ctx, idx, name, parent)
+		if err != nil {
+			return nil, err
+		}
+		if node == nil {
+			return nil, errNilNode
+		}
+		return node, nil
+	})
+	switch err {
+	case nil:
+		node = item.(*Node)
+	case errNilNode:
+		err = nil
+	}
+	return
+}
+
 func (t *Tree) nodeFullName(name string, parent *Node) string {
 	if parent != nil {
 		name = parent.Name + "." + name
 	}
 	return name
-}
-
-func (t *Tree) Nodes(idx int, name string) *Node {
-	if len(t.nodes) <= idx {
-		return nil
-	}
-	item := t.nodes[idx].Get(name)
-	if item != nil && !item.Expired() {
-		return item.Value().(*Node)
-	}
-	return nil
 }
 
 func (t *Tree) createNode(ctx context.Context, idx int, name string, parent *Node) (node *Node, err error) {
@@ -133,26 +149,12 @@ func (t *Tree) createNode(ctx context.Context, idx int, name string, parent *Nod
 	node.Name = name
 	node.Tree = t
 	node.Level = idx
-	if parent != nil {
-		parent.Childs.PutIfAbsent(name, struct{}{})
-	}
-	return
-}
-
-// Simulate will create a temp node from tree
-func (t *Tree) Simulate(ctx context.Context) (node *Node, err error) {
-	var parent *Node
-	for idx, filter := range t.filters {
-		name := t.nodeFullName(filter.Name(ctx), parent)
-		node, err = t.createNode(ctx, idx, name, parent)
-		if node == nil {
-			return
-		}
-		parent = node
-	}
 	return
 }
 
 func NewTree(cfg *Config) *Tree {
-	return &Tree{Config: cfg}
+	return &Tree{
+		Config: cfg,
+		roots:  ccache.New(ccache.Configure().MaxSize(cfg.MaxSize())),
+	}
 }
