--- server/service/event/service_alias_event_handler.go
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package event
-
-import (
-	"github.com/apache/incubator-servicecomb-service-center/server/core/backend"
-	pb "github.com/apache/incubator-servicecomb-service-center/server/core/proto"
-	"github.com/apache/incubator-servicecomb-service-center/server/service/cache"
-)
-
-type ServiceAliasEventHandler struct {
-}
-
-func (h *ServiceAliasEventHandler) Type() backend.StoreType {
-	return backend.SERVICE_ALIAS
-}
-
-func (h *ServiceAliasEventHandler) OnEvent(evt backend.KvEvent) {
-	switch evt.Type {
-	case pb.EVT_DELETE:
-		providerKey := backend.GetInfoFromSvcIndexKV(evt.KV)
-		cache.FindInstances.Remove(providerKey)
-	}
-}
-
-func NewServiceAliasEventHandler() *ServiceAliasEventHandler {
-	return &ServiceAliasEventHandler{}
-}
