+++ server/service/event/tag_event_handler.go
@@ -22,6 +22,7 @@ import (
 	"github.com/apache/incubator-servicecomb-service-center/pkg/util"
 	"github.com/apache/incubator-servicecomb-service-center/server/core/backend"
 	pb "github.com/apache/incubator-servicecomb-service-center/server/core/proto"
+	"github.com/apache/incubator-servicecomb-service-center/server/service/cache"
 	nf "github.com/apache/incubator-servicecomb-service-center/server/service/notification"
 	serviceUtil "github.com/apache/incubator-servicecomb-service-center/server/service/util"
 	"golang.org/x/net/context"
@@ -50,23 +51,29 @@ func (apt *TagsChangedTask) Err() error {
 }
 
 func (apt *TagsChangedTask) publish(ctx context.Context, domainProject, consumerId string, rev int64) error {
-	consumer, err := serviceUtil.GetServiceInCache(ctx, domainProject, consumerId)
+	ctx = util.SetContext(ctx, serviceUtil.CTX_CACHEONLY, "1")
+
+	consumer, err := serviceUtil.GetService(ctx, domainProject, consumerId)
 	if err != nil {
-		util.Logger().Errorf(err, "get comsumer for publish event %s failed", consumerId)
+		util.Logger().Errorf(err, "get consumer for publish event %s failed", consumerId)
 		return err
 	}
 	if consumer == nil {
 		util.Logger().Errorf(nil, "service not exist, %s", consumerId)
 		return fmt.Errorf("service not exist, %s", consumerId)
 	}
-	providerIds, err := serviceUtil.GetProvidersInCache(ctx, domainProject, consumer)
+
+	serviceKey := pb.MicroServiceToKey(domainProject, consumer)
+	cache.FindInstances.Remove(serviceKey)
+
+	providerIds, err := serviceUtil.GetProviderIds(ctx, domainProject, consumer)
 	if err != nil {
 		util.Logger().Errorf(err, "get provider services by consumer %s failed", consumerId)
 		return err
 	}
 
 	for _, providerId := range providerIds {
-		provider, err := serviceUtil.GetServiceInCache(ctx, domainProject, providerId)
+		provider, err := serviceUtil.GetService(ctx, domainProject, providerId)
 		if provider == nil {
 			util.Logger().Warnf(err, "get service %s file failed", providerId)
 			continue
@@ -94,11 +101,11 @@ func (h *TagEventHandler) OnEvent(evt backend.KvEvent) {
 	consumerId, domainProject := backend.GetInfoFromTagKV(evt.KV)
 
 	if nf.GetNotifyService().Closed() {
-		util.Logger().Warnf(nil, "caught service %s tags [%s] event, but notify service is closed",
-			consumerId, action)
+		util.Logger().Warnf(nil, "caught [%s] service tags event %s/%s, but notify service is closed",
+			action, consumerId, evt.KV.Value)
 		return
 	}
-	util.Logger().Infof("caught service %s tags [%s] event", consumerId, action)
+	util.Logger().Infof("caught [%s] service tags event %s/%s", action, consumerId, evt.KV.Value)
 
 	async.Service().Add(context.Background(),
 		NewTagsChangedAsyncTask(domainProject, consumerId, evt.Revision))
