+++ server/service/notification/stream.go
@@ -24,14 +24,15 @@ import (
 )
 
 func HandleWatchJob(watcher *ListWatcher, stream pb.ServiceInstanceCtrl_WatchServer) (err error) {
+	timer := time.NewTimer(DEFAULT_SEND_TIMEOUT)
+	defer timer.Stop()
 	for {
-		timer := time.NewTimer(DEFAULT_SEND_TIMEOUT)
 		select {
 		case <-timer.C:
+			timer.Reset(DEFAULT_SEND_TIMEOUT)
+
 			// TODO grpc 长连接心跳？
 		case job := <-watcher.Job:
-			timer.Stop()
-
 			if job == nil {
 				err = errors.New("channel is closed")
 				util.Logger().Errorf(err, "watcher caught an exception, subject: %s, group: %s",
@@ -49,6 +50,8 @@ func HandleWatchJob(watcher *ListWatcher, stream pb.ServiceInstanceCtrl_WatchSer
 				watcher.SetError(err)
 				return
 			}
+
+			util.ResetTimer(timer, DEFAULT_SEND_TIMEOUT)
 		}
 	}
 }
