+++ server/service/util/dependency_test.go
@@ -27,42 +27,6 @@ func TestDeleteDependencyForService(t *testing.T) {
 	if err != nil {
 		t.Fatalf(`DeleteDependencyForDeleteService failed`)
 	}
-
-	_, err = updateProviderDependencyRuleUtil(
-		&proto.MicroServiceDependency{
-			Dependency: []*proto.MicroServiceKey{
-				{AppId: "a"},
-			},
-		},
-		&proto.MicroServiceKey{
-			AppId: "a",
-		}, "")
-	if err != nil {
-		t.Fatalf(`deleteDependencyRuleUtil with the same deps failed`)
-	}
-
-	_, err = updateProviderDependencyRuleUtil(
-		&proto.MicroServiceDependency{
-			Dependency: []*proto.MicroServiceKey{
-				{AppId: "b"},
-			},
-		},
-		&proto.MicroServiceKey{
-			AppId: "a",
-		}, "")
-	if err != nil {
-		t.Fatalf(`deleteDependencyRuleUtil failed`)
-	}
-
-	_, err = deleteConsumerDepOfProviderRule(context.Background(), "", &proto.MicroServiceKey{}, &proto.MicroServiceKey{})
-	if err == nil {
-		t.Fatalf(`deleteConsumerDepOfProviderRule failed`)
-	}
-
-	_, err = deleteDepRuleUtil("", &proto.MicroServiceDependency{}, &proto.MicroServiceKey{})
-	if err != nil {
-		t.Fatalf(`deleteDepRuleUtil failed`)
-	}
 }
 
 func TestTransferToMicroServiceDependency(t *testing.T) {
@@ -108,11 +72,6 @@ func TestCreateDependencyRule(t *testing.T) {
 		t.Fatalf(`AddServiceVersionRule failed`)
 	}
 
-	_, err = addDepRuleUtil("", &proto.MicroServiceDependency{}, &proto.MicroServiceKey{})
-	if err != nil {
-		t.Fatalf(`addDepRuleUtil failed`)
-	}
-
 	b, err := containServiceDependency([]*proto.MicroServiceKey{
 		{AppId: "a"},
 	}, &proto.MicroServiceKey{
@@ -147,33 +106,6 @@ func TestCreateDependencyRule(t *testing.T) {
 	})
 	if !ok {
 		t.Fatalf(`diffServiceVersion failed`)
-	}
-
-	ok = isDependencyAll(&proto.MicroServiceDependency{})
-	if ok {
-		t.Fatalf(`isDependencyAll not * failed`)
-	}
-
-	ok = isDependencyAll(&proto.MicroServiceDependency{
-		Dependency: []*proto.MicroServiceKey{
-			{
-				ServiceName: "*",
-			},
-		},
-	})
-	if !ok {
-		t.Fatalf(`isDependencyAll * failed`)
-	}
-
-	ok = isExist([]*proto.MicroServiceKey{
-		{
-			ServiceName: "*",
-		},
-	}, &proto.MicroServiceKey{
-		ServiceName: "*",
-	})
-	if !ok {
-		t.Fatalf(`isExist failed`)
 	}
 }
 
@@ -249,11 +181,6 @@ func TestServiceDependencyRuleExist(t *testing.T) {
 }
 
 func TestUpdateServiceForAddDependency(t *testing.T) {
-	_, _, err := updateDepRuleUtil("", &proto.MicroServiceDependency{}, &proto.MicroServiceKey{})
-	if err != nil {
-		t.Fatalf(`updateDepRuleUtil failed`)
-	}
-
 	old := isNeedUpdate([]*proto.MicroServiceKey{
 		{
 			AppId:       "a",
@@ -267,33 +194,6 @@ func TestUpdateServiceForAddDependency(t *testing.T) {
 	})
 	if old == nil {
 		t.Fatalf(`isNeedUpdate failed`)
-	}
-}
-
-func TestFilter(t *testing.T) {
-	_, _, err := getConsumerIdsWithFilter(context.Background(), "", &proto.MicroService{}, noFilter)
-	if err == nil {
-		t.Fatalf(`getConsumerIdsWithFilter failed`)
-	}
-
-	_, _, err = filterConsumerIds(context.Background(), []string{}, noFilter)
-	if err != nil {
-		t.Fatalf(`filterConsumerIds invalid failed`)
-	}
-
-	_, _, err = filterConsumerIds(context.Background(), []string{"a"}, noFilter)
-	if err != nil {
-		t.Fatalf(`filterConsumerIds invalid failed`)
-	}
-
-	rf := RuleFilter{
-		DomainProject: "",
-		Provider:      &proto.MicroService{},
-		ProviderRules: []*proto.ServiceRule{},
-	}
-	_, _, err = filterConsumerIds(context.Background(), []string{"a"}, rf.Filter)
-	if err == nil {
-		t.Fatalf(`filterConsumerIds invalid failed`)
 	}
 }
 
@@ -355,6 +255,10 @@ func TestDependency(t *testing.T) {
 	if err == nil {
 		t.Fatalf(`DependencyRelation_getDependencyConsumersOfProvider failed`)
 	}
+	_, err = dr.GetDependencyProviders()
+	if err == nil {
+		t.Fatalf(`DependencyRelation_GetDependencyProviders failed`)
+	}
 }
 
 func TestDependencyRelationFilterOpt(t *testing.T) {
@@ -364,5 +268,12 @@ func TestDependencyRelationFilterOpt(t *testing.T) {
 	)
 	if !op.NonSelf || !op.SameDomainProject {
 		t.Fatalf(`toDependencyRelationFilterOpt failed`)
+	}
+}
+
+func TestGetConsumerIdsWithFilter(t *testing.T) {
+	_, _, err := getConsumerIdsWithFilter(context.Background(), "", &proto.MicroService{}, nil)
+	if err == nil {
+		t.Fatalf(`TestGetConsumerIdsWithFilter failed`)
 	}
 }
